requirements.txt
chromedriver==2.24.1
coverage==4.2
django-choices==1.5.0
dj-database-url==0.4.1
Django==1.10.3
django-bootstrap3==7.1.0
django-ckeditor==5.1.1
django-contrib-comments==1.7.3
django-datetime-widget==0.9.3
django-jsonfield==1.0.1
django-rq==0.9.4
gunicorn==19.6.0
mock==2.0.0
page-objects==1.1.0
psycopg2==2.6.2
pytz==2016.7
redis==2.10.5
rq==0.7.1
selenium==2.53.6
social-auth-app-django==1.1.0
sparkpost==1.3.2
wheel==0.24.0
whitenoise==3.2.2
sunlight==1.2.9
geopy==1.11.0
https://github.com/awatts/django-activity-stream/archive/ad4211199546fcbfe43e8f936c5d1265691c1ef4.zip
factory_boy==2.8.1
